css/css-flexbox/parsing/align-self-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Flexible Box Layout Module Level 1: parsing align-self with valid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-flexbox/#propdef-align-self">
<meta name="assert" content="align-self supports the full grammar 'auto | flex-start | flex-end | center | baseline | stretch'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("align-self", "auto");
test_valid_value("align-self", "flex-start");
test_valid_value("align-self", "flex-end");
test_valid_value("align-self", "center");
test_valid_value("align-self", "baseline");
test_valid_value("align-self", "stretch");
</script>
</body>
</html>
